Change type for endpoints returning issues list.
This change corrects a problem where it was not possible to get a list of
customized issues because it was not possible to extends 'Issues' itself. Now
with a 'List<? extends Issue>' instead of 'Issues', anyone can create their
own 'Issues' class when they want to use specific 'Issue0 class containing
customized fields.